scripts/requirements.txt
# PSC-Graph 项目完整依赖清单
# 更新时间: 2025-11-17
# 遵循CLAUDE.md规范：锁定版本号确保可重现

# ============================================================
# 核心依赖（必须）
# ============================================================

# HTTP客户端与网络请求
requests==2.32.5

# HTML解析
beautifulsoup4==4.14.2
lxml==6.0.2

# PDF解析
pdfplumber==0.11.8

# 数据处理
pandas==2.3.3
numpy==2.3.5
scipy==1.16.3

# 配置文件解析
PyYAML==6.0

# ============================================================
# RAG索引与检索依赖
# ============================================================

# 中文分词
jieba==0.42.1

# 机器学习与向量化
scikit-learn==1.7.2

# 向量检索
faiss-cpu==1.13.0

# 语义向量模型
sentence-transformers==5.1.2
torch==2.9.1

# ============================================================
# 图学习依赖
# ============================================================

# PyTorch Geometric（异质图学习）
torch-geometric==2.7.0

# PyG扩展（可选，用于加速，可能需要编译）
# torch-scatter  # 需要与torch版本匹配
# torch-sparse   # 需要与torch版本匹配

# ============================================================
# 标注验证依赖
# ============================================================

# JSON Schema验证
jsonschema==4.25.1

# ============================================================
# 可选依赖（增强功能）
# ============================================================

# 高级PDF解析（需要Java和ghostscript）
# camelot-py[cv]==0.11.0

# PDF转图片（需要poppler）
# pdf2image==1.16.3

# OCR引擎（需要tesseract）
# pytesseract==0.3.10

# ============================================================
# 安装说明
# ============================================================
#
# 1. 基础安装（数据爬取 + 标注验证）：
#    pip install -r requirements.txt
#
# 2. 系统依赖（Ubuntu/Debian）：
#    apt-get install python3-jieba  # 推荐使用系统包
#
# 3. CUDA支持（可选，用于GPU加速）：
#    - torch会自动安装CUDA 12.8支持
#    - 如无GPU，torch仍可在CPU上运行
#
# 4. 验证安装：
#    python3 -c "import sentence_transformers, faiss, jieba; print('✓ 所有依赖已安装')"
#
